Fixed difference between the primary key types of a flight domain model and the type specified in the JPA repository
Adjusted FlightSerivce and its tests to use the correct type of Long instead of Int

+++ src/main/java/com/aviobook/schedule/repository/FlightRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface FlightRepository extends JpaRepository<Flight, Integer>, JpaSpecificationExecutor<Flight> {
+public interface FlightRepository extends JpaRepository<Flight, Long>, JpaSpecificationExecutor<Flight> {
     boolean existsByNumber(String number);
 }
